Finish flights-contracts: move snapshot builder via catalog snapshot carve
Carve the pure PricingBasis + CaptureSnapshotInput shapes out of the catalog
runtime (services/snapshot-service.ts, snapshot/schema.ts) into
@voyantjs/catalog-contracts/snapshot. catalog re-exports them from their
original paths, so all consumers (accommodations/charters/cruises/extras/
products, catalog-ui) are unchanged.

Move flights/src/snapshot.ts into @voyantjs/flights-contracts, importing the
snapshot types from catalog-contracts; the flights runtime re-exports it. The
flight snapshot builder no longer needs the catalog runtime.

Resolves #1449.
