+++ packages/catalog-contracts/package.json
@@ -15,7 +15,8 @@
     "./adapter/channel-push-contracts": "./src/adapter/channel-push-contracts.ts",
     "./adapter/provider-contracts": "./src/adapter/provider-contracts.ts",
     "./booking-engine/contracts": "./src/booking-engine/contracts.ts",
-    "./booking-engine/promotions-contract": "./src/booking-engine/promotions-contract.ts"
+    "./booking-engine/promotions-contract": "./src/booking-engine/promotions-contract.ts",
+    "./snapshot": "./src/snapshot.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
@@ -90,6 +91,11 @@
         "types": "./dist/booking-engine/promotions-contract.d.ts",
         "import": "./dist/booking-engine/promotions-contract.js",
         "default": "./dist/booking-engine/promotions-contract.js"
+      },
+      "./snapshot": {
+        "types": "./dist/snapshot.d.ts",
+        "import": "./dist/snapshot.js",
+        "default": "./dist/snapshot.js"
       }
     },
     "main": "./dist/index.js",
