+++ packages/catalog-contracts/src/index.ts
@@ -6,3 +6,4 @@ export * from "./content.js"
 export * from "./contract.js"
 export * from "./drift/events.js"
 export * from "./provenance.js"
+export * from "./snapshot.js"
