+++ packages/flights-contracts/src/index.ts
@@ -4,3 +4,4 @@ export * from "./contract/schemas.js"
 export * from "./contract/types.js"
 export * from "./reference/contract.js"
 export * from "./reference/static-bundle.js"
+export * from "./snapshot.js"
